+++ app/src/org/commcare/activities/DispatchActivity.java
@@ -385,24 +385,16 @@ public class DispatchActivity extends AppCompatActivity {
     }
 
     private void launchHomeScreen() {
-        Intent i;
-        if (useRootMenuHomeActivity()) {
-            i = new Intent(this, RootMenuHomeActivity.class);
-            // Since we are entering a menu list, the session state will expect this later
-            HomeScreenBaseActivity.addPendingDataExtra(
-                    i,
-                    CommCareApplication.instance().getCurrentSessionWrapper().getSession()
-            );
-        } else {
-            i = new Intent(this, StandardHomeActivity.class);
-        }
-        i.putExtra(START_FROM_LOGIN, startFromLogin);
-        i.putExtra(LoginActivity.LOGIN_MODE, lastLoginMode);
-        i.putExtra(LoginActivity.MANUAL_SWITCH_TO_PW_MODE, userManuallyEnteredPasswordMode);
-        i.putExtra(PERSONALID_MANAGED_LOGIN, personalIdManagedLogin);
+        Intent intent = HomeScreenBaseActivity.buildHomeIntent(
+                this,
+                lastLoginMode,
+                startFromLogin,
+                userManuallyEnteredPasswordMode,
+                personalIdManagedLogin
+        );
         startFromLogin = false;
         clearSessionEndpointIntentExtras();
-        startActivityForResult(i, HOME_SCREEN);
+        startActivityForResult(intent, HOME_SCREEN);
     }
 
     public static boolean useRootMenuHomeActivity() {